% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_annual_estimates}
\alias{load_annual_estimates}
\title{Load Methane Dataset 1}
\usage{
load_annual_estimates()
}
\value{
A data frame containing Soil Chamber data.
}
\description{
Returns the first Soil Chamber dataset.
}
