% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_beringian_wetlands}
\alias{load_beringian_wetlands}
\title{Load Methane Dataset 2}
\usage{
load_beringian_wetlands()
}
\value{
A data frame containing Soil Chamber data.
}
\description{
Returns the second Soil Chamber dataset.
}
