% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_soilchamber_1}
\alias{load_soilchamber_1}
\title{Load Methane Dataset 3}
\usage{
load_soilchamber_1()
}
\value{
A data frame containing CH4 flux data.
}
\description{
Returns the Beringian Coastal Wetlands dataset.
}
