% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_soilchamber_2}
\alias{load_soilchamber_2}
\title{Load Methane Dataset 4}
\usage{
load_soilchamber_2()
}
\value{
A data frame containing annual CH4 estimates.
}
\description{
Returns the Annual CH4 estimates dataset.
}
